+++ spec/spec_helper.rb
@@ -1,9 +1,18 @@
 # frozen_string_literal: true
 ENV['RACK_ENV'] = ENV['ENVIRONMENT'] ||= 'test'
 require 'simplecov'
+require 'simplecov-json'
 if ENV['TRAVIS']
   require 'codecov'
-  SimpleCov.formatter = SimpleCov::Formatter::Codecov
+  SimpleCov.formatters = [
+    SimpleCov::Formatter::JSONFormatter,
+    SimpleCov::Formatter::Codecov
+  ]
+else
+  SimpleCov.formatters = [
+    SimpleCov::Formatter::HTMLFormatter,
+    SimpleCov::Formatter::JSONFormatter
+  ]
 end
 SimpleCov.start do
   add_filter '/spec/'