+++ lib/request_handler/base.rb
@@ -135,7 +135,9 @@ module RequestHandler
     def params
       raise MissingArgumentError, params: 'is missing' if request.params.nil?
       raise ExternalArgumentError, params: 'must be a Hash' unless request.params.is_a?(Hash)
-      @params ||= Helper.deep_transform_keys_in_object(request.params) { |k| k.to_s.gsub('.', ::RequestHandler.separator) }
+      @params ||= Helper.deep_transform_keys_in_object(request.params) do |k|
+        k.to_s.gsub('.', ::RequestHandler.separator)
+      end
     end
 
     def config