+++ spec/spec_helper.rb
@@ -3,7 +3,7 @@ ENV['RACK_ENV'] = ENV['ENVIRONMENT'] ||= 'test'
 require 'simplecov'
 if ENV['TRAVIS']
   require 'codecov'
-  SimpleCov.formatter = SimpleCov::Formatter::Codecov
+  SimpleCov.formatter SimpleCov::Formatter::Codecov
 end
 SimpleCov.start do
   add_filter '/spec/'