+++ lib/request_handler/include_option_parser.rb
@@ -14,7 +14,7 @@ module RequestHandler
       options.map do |option|
         option.gsub!('.', ::RequestHandler.separator)
         begin
-          allowed_options_type&.call(option).to_sym
+          allowed_options_type.call(option).to_sym
         rescue Dry::Types::ConstraintError
           raise OptionNotAllowedError, option.to_sym => 'is not an allowed include option'
         end