+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - raise an error if mandatory options are missing in the handler configuration
 - Transform param keys to string before substitution
 - make gem compatible with ruby 2.1
+- add danger
 
 ## 0.9.1