+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 - raise an error if mandatory options are missing in the handler configuration
 - Transform param keys to string before substitution
+- raise an ExternalArgumentError if the body does not contain data
 
 ## 0.9.1